whatever

+++ setup_project.sh
@@ -12,10 +12,10 @@ mkdir -p "$project/Helpers"
 mkdir -p "$project/reports"
 
 # Moving existing source files into correct structure
-mv attendance_checker.py "$project/"
-mv assets.csv "$project/Helpers/"
-mv config.json "$project/Helpers/"
-mv reports.log "$project/reports/"
+[ -f "attendance_checker.py" ] && mv attendance_checker.py "$project/"
+[ -f "assets.csv" ] && mv assets.csv "$project/Helpers/"
+[ -f "config.json" ] && mv config.json "$project/Helpers/"
+[ -f "reports.log" ] && mv reports.log "$project/reports/"
 
 #Archiving and cleaning Interrupted project
 trap 'echo "Interrupted! Archiving project..."; tar -czf "${project}_archive.tar.gz" "${project}";
@@ -25,17 +25,20 @@ rm -rf "$project"; exit' SIGINT
 read -p "Do you want to update thresholds? (yes/no): " choice
 
 if [ "$choice" = "yes" ]; then
+	config_file=$project/Helpers/config.json
 
     # Get new values from user
-    read -p "Enter new warning threshold: " warning
-    read -p "Enter new failure threshold: " failure
+        read -p "Enter new warning threshold: " warning
+        read -p "Enter new failure threshold: " failure
 
     # Update config.json safely
-    sed -i "s/\"warning\": [0-9]*/\"warning\": $warning/" $project/Helpers/config.json
-    sed -i "s/\"failure\": [0-9]*/\"failure\": $failure/" $project/Helpers/config.json
+        if [ -f "$config_file" ]; then
+		sed -i "s/\"warning\": [0-9]*/\"warning\": $warning/" "$config_file"
+		sed -i "s/\"failure\": [0-9]*/\"failure\": $failure/" "$config_file"
+        
 
 else
-    echo "Keeping default threshold values"
+    echo "Config file not found, skipping update"
 fi
 
 # Check if Python3 is installed
